+++ pyproject.toml
@@ -49,8 +49,7 @@ dependencies = [
     "timm==0.9.2",
     "wandb>=0.13.9",
     "xarray>=0.20.2",
-    "rasterio>=1.3.7",
-    "xesmf>=0.7.1"
+    "rasterio>=1.3.7"
 ]
 
 [project.optional-dependencies]
